requirements.txt
celery===4.4.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
python-decouple==3.6
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect==1.1.0
cssutils==2.3.0
defusedxml==0.7.1
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
dnspython==2.2.0
drf-yasg==1.20.0
ecdsa==0.17.0
email-validator==1.1.3
emails==0.5.15
fastapi==0.70.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.9.0
httptools==0.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
inflect==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==3.0.37
lxml==4.7.1
Mako==1.1.6
MarkupSafe==2.0.1
more-itertools==8.12.0
oauthlib==3.2.0
packaging==21.3
passlib==1.7.4
password==0.2
Pillow==9.0.0
pluggy==0.13.1
premailer==3.10.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==5.4.3
python-dateutil==2.8.2
python-decouple==3.5
python-jose==3.3.0
python-multipart==0.0.5
python-social-auth==0.2.12
python3-openid==3.2.0
pytz==2021.3
raven==6.10.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
starlette==0.16.0
tenacity==6.3.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.11.8
uvloop==0.16.0
vine==1.3.0
wcwidth==0.2.5
websockets==8.1
zipp==3.7.0
redis==4.1.2

